Fixed Menus rendering only as shadows when drawing to a parent with negative Y-coordinates relative to its canvas.
The Menu::Layout function alters the height of the menu to fit the canvas, but uses only the coordinate relative to its parent to check if it fits, instead of the bounds relative to the canvas.

The problem was resolved by using LocalToCanvas() in place of Y() for comparing the position of the menu relative to the canvas.